requirements.txt
Flask
Flask-PyMongo
Flask-Login
Flask-WTF
Flask-Mail
python-dateutil
python-dotenv
gunicorn